gpu: nvgpu: Fix overflow of alloc length
Fix an issue where large ( > 4GB) allocations were not being computed
correctly. The two fields, pages and page_size, were both 32 bits so
when multiplied they easily overflowed. Simple fix is to cast them to
64 bits before multiplying them.

Reviewed-on: http://git-master/r/747429
Reviewed-on: http://git-master/r/771148

+++ drivers/gpu/nvgpu/gk20a/mm_gk20a.c
@@ -2816,7 +2816,9 @@ int gk20a_vm_alloc_space(struct gk20a_as_share *as_share,
 						 (u64)args->pages *
 						 (u64)args->page_size);
 	else
-		vaddr_start = gk20a_balloc(vma, args->pages * args->page_size);
+		vaddr_start = gk20a_balloc(vma,
+					   (u64)args->pages *
+					   (u64)args->page_size);
 
 	if (!vaddr_start) {
 		kfree(va_node);
